requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=- requirements.in
#
click==8.3.2
    # via -r requirements.in
coverage[toml]==7.13.5
    # via pytest-cov
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via pytest-html
markupsafe==3.0.3
    # via jinja2
packaging==26.0
    # via pytest
piexif==1.1.3
    # via -r requirements.in
pillow==12.2.0
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pygments==2.20.0
    # via pytest
pytest==9.0.3
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-html
    #   pytest-metadata
    #   pytest-mock
    #   pytest-random-order
pytest-cov==7.1.0
    # via -r requirements.in
pytest-html==4.2.0
    # via -r requirements.in
pytest-metadata==3.1.1
    # via pytest-html
pytest-mock==3.15.1
    # via -r requirements.in
pytest-random-order==1.2.0
    # via -r requirements.in
